#0: split ttnn normalization .hpp/.cpp and cleanup compute_kernel_config and sharding_utils
