+++ ttnn/cpp/ttnn/deprecated/tt_dnn/op_library/CMakeLists.txt
@@ -2,6 +2,8 @@
 # We do not use GLOB here since build system won't be able to pick up changes to the file list generated
 set(TT_DNN_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/auto_format.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/sharding_utilities.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/compute_kernel_config.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/sharded/sharded_op.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/sharded/multi_core/sharded_op_multi_core.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bcast/bcast_op.cpp
