+++ ttnn/cpp/ttnn/operations/normalization/groupnorm/groupnorm_pybind.hpp
@@ -4,39 +4,10 @@
 
 #pragma once
 
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-
-#include "ttnn/cpp/pybind11/decorators.hpp"
-#include "groupnorm.hpp"
-
-namespace py = pybind11;
+#include "pybind11/pybind_fwd.hpp"
 
 namespace ttnn::operations::normalization::detail {
 
-void bind_normalization_group_norm_operation(py::module& module) {
-
-    ttnn::bind_registered_operation(
-        module,
-        ttnn::group_norm,
-        R"doc(group_norm(input_tensor: ttnn.Tensor, *, num_groups: int, epsilon: float = 1e-12, weight: Optional[ttnn.Tensor] = None, bias: Optional[ttnn.Tensor] = None) -> ttnn.Tensor
-          Compute group_norm over :attr:`input_tensor`.
-        )doc",
-        ttnn::pybind_arguments_t{
-            py::arg("input_tensor"),
-            py::kw_only(),
-            py::arg("num_groups"),
-            py::arg("epsilon") = 1e-12,
-            py::arg("input_mask") = std::nullopt,
-            py::arg("weight") = std::nullopt,
-            py::arg("bias") = std::nullopt,
-            py::arg("memory_config") = std::nullopt,
-            py::arg("dtype") = std::nullopt,
-            py::arg("core_grid") = std::nullopt,
-            py::arg("inplace") = true,
-            py::arg("output_layout") = std::nullopt
-        }
-    );
-}
+void bind_normalization_group_norm(pybind11::module& module);
 
 }  // namespace ttnn::operations::normalization::detail
