+++ ttnn/cpp/ttnn/operations/normalization/normalization_pybind.hpp
@@ -17,17 +17,10 @@
 namespace ttnn::operations::normalization {
 
 void py_module(py::module& module) {
-
-    detail::bind_normalization_softmax_program_config_operation(module);
-    detail::bind_normalization_softmax_operation(module);
-    detail::bind_normalization_scale_mask_softmax_operation(module);
-    detail::bind_normalization_softmax_in_place_operation(module);
-    detail::bind_normalization_scale_mask_softmax_in_place_operation(module);
-    detail::bind_normalization_scale_causal_mask_hw_dims_softmax_in_place_operation(module);
-    detail::bind_normalization_layernorm_program_config_operation(module);
-    detail::bind_normalization_layer_norm_operation(module);
-    detail::bind_normalization_rms_norm_operation(module);
-    detail::bind_normalization_group_norm_operation(module);
+    detail::bind_normalization_softmax(module);
+    detail::bind_normalization_layernorm(module);
+    detail::bind_normalization_rms_norm(module);
+    detail::bind_normalization_group_norm(module);
 }
 
 }  // namespace ttnn::operations::normalization
