+++ ttnn/cpp/ttnn/operations/normalization/softmax/device/softmax_types.hpp
@@ -1,0 +1,25 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "common/core_coord.h"
+
+namespace ttnn::operations::normalization {
+
+struct SoftmaxDefaultProgramConfig{
+};
+struct SoftmaxShardedMultiCoreProgramConfig {
+    CoreCoord compute_with_storage_grid_size;
+    std::size_t subblock_w;
+    std::size_t block_h;
+    std::size_t block_w;
+};
+
+using SoftmaxProgramConfig = std::variant<
+    SoftmaxDefaultProgramConfig,
+    SoftmaxShardedMultiCoreProgramConfig
+>;
+
+}  // namespace ttnn::operations::normalization
