+++ ttnn/cpp/ttnn/operations/normalization/softmax/softmax.hpp
@@ -4,9 +4,10 @@
 
 #pragma once
 
-#include "ttnn/deprecated/tt_dnn/op_library/moreh_softmax/moreh_softmax_op.hpp"
-#include "device/softmax_op.hpp"
-#include "ttnn/operations/core/core.hpp"
+#include "ttnn/decorators.hpp"
+#include "ttnn/deprecated/tt_dnn/op_library/compute_kernel_config.hpp"
+
+#include "device/softmax_types.hpp"
 
 namespace ttnn {
 namespace operations::normalization {
@@ -17,25 +18,7 @@ struct ExecuteSoftmax {
         const ttnn::Tensor& input_tensor,
         const int dim_arg,
         const std::optional<ttnn::MemoryConfig>& memory_config = std::nullopt,
-        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt) {
-        auto input_shape = input_tensor.get_shape();
-        auto rank = input_shape.size();
-        auto dim = dim_arg;
-        if (dim < 0) {
-            dim = rank + dim;
-        }
-
-        auto input_tensor_4D = ttnn::unsqueeze_to_4D(input_tensor);
-        if (dim == rank - 1) {
-            auto output_tensor = ttnn::operations::normalization::softmax(
-                input_tensor_4D, memory_config.value_or(input_tensor.memory_config()), compute_kernel_config);
-            return ttnn::reshape(output_tensor, input_shape);
-        } else {
-            auto dim_4D = dim + 4 - rank;
-            auto output_tensor = tt::operations::primary::moreh_softmax(input_tensor_4D, dim_4D);
-            return ttnn::reshape(output_tensor, input_shape);
-        }
-    }
+        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt);
 };
 
 struct ExecuteScaleMaskSoftmax {
@@ -46,14 +29,7 @@ struct ExecuteScaleMaskSoftmax {
         const std::optional<const Tensor> mask = std::nullopt,
         const std::optional<ttnn::MemoryConfig>& memory_config = std::nullopt,
         const bool is_causal_mask = false,
-        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt) {
-        auto input_shape = input_tensor.get_shape();
-
-        auto input_tensor_4D = ttnn::unsqueeze_to_4D(input_tensor);
-        auto output_tensor =
-            ttnn::operations::normalization::scale_mask_softmax(input_tensor_4D, scale, mask, memory_config.value_or(input_tensor.memory_config()), is_causal_mask, compute_kernel_config);
-        return ttnn::reshape(output_tensor, input_shape);
-    }
+        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt);
 };
 
 struct ExecuteSoftmaxInPlace {
@@ -62,14 +38,7 @@ struct ExecuteSoftmaxInPlace {
     static ttnn::Tensor invoke(
         const ttnn::Tensor& input_tensor,
         const SoftmaxProgramConfig& program_config = SoftmaxDefaultProgramConfig{},
-        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt) {
-        auto input_shape = input_tensor.get_shape();
-
-        auto input_tensor_4D = ttnn::unsqueeze_to_4D(input_tensor);
-        auto output_tensor =
-            ttnn::operations::normalization::softmax_in_place(input_tensor_4D, program_config, compute_kernel_config);
-        return ttnn::reshape(output_tensor, input_shape);
-    }
+        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt);
 };
 
 struct ExecuteScaleMaskSoftmaxInPlace {
@@ -81,14 +50,7 @@ struct ExecuteScaleMaskSoftmaxInPlace {
         const std::optional<const Tensor> mask = std::nullopt,
         const SoftmaxProgramConfig& program_config = SoftmaxDefaultProgramConfig{},
         const bool is_causal_mask = false,
-        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt) {
-        auto input_shape = input_tensor.get_shape();
-
-        auto input_tensor_4D = ttnn::unsqueeze_to_4D(input_tensor);
-        auto output_tensor =
-            ttnn::operations::normalization::scale_mask_softmax_in_place(input_tensor_4D, scale, mask, program_config, is_causal_mask, compute_kernel_config);
-        return ttnn::reshape(output_tensor, input_shape);
-    }
+        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt);
 };
 
 struct ExecuteScaleCausalMaskHWSoftmaxInPlace {
@@ -99,14 +61,7 @@ struct ExecuteScaleCausalMaskHWSoftmaxInPlace {
         const std::optional<float> scale = std::nullopt,
         const std::optional<const Tensor> mask = std::nullopt,
         const SoftmaxProgramConfig& program_config = SoftmaxDefaultProgramConfig{},
-        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt) {
-        auto input_shape = input_tensor.get_shape();
-
-        auto input_tensor_4D = ttnn::unsqueeze_to_4D(input_tensor);
-        auto output_tensor =
-            ttnn::operations::normalization::scale_causal_mask_hw_dims_softmax_in_place(input_tensor_4D, scale, mask, program_config, compute_kernel_config);
-        return ttnn::reshape(output_tensor, input_shape);
-    }
+        const std::optional<const DeviceComputeKernelConfig> compute_kernel_config = std::nullopt);
 };
 
 }  // namespace operations::normalization
